Add projectsupport function modify project model
After creating project, it would jump to editing project
page in order to add the projectsupport. One project
correspone to one projectsupport.

Add status for project model to display weather project
is set projectsupport or not. If not, project wouldn't
show in the index page.

Add function for searching by category.

#20211121001